Movement behaviour logic in progress.

+++ Assets/Scripts/Configs/Entity/EntityConfig.cs
@@ -13,7 +13,7 @@ namespace Configs.Entity
         [Header("Movement Parameters"), Range(0f, 10f)]
         public float Speed;
 
-        public float2 StartDirection = new(0, -1);
+        public int2 StartDirection = new(0, -1);
 
         [Header("Health Parameters"), Range(0, 5)]
         public int Health;
